zuul: remove wazo-ansible-uc-archive job
Why:

* Since we changed the Ansible role according to the new behavior of
wazo-calld, it is not suitable to older versions of wazo-calld. The
point of the Zuul job is to check whether the master Ansible roles can
still install an archive version, but in this case the two are
incompatible. We should only run this check on the latest archive repo,
created at release time.